Refactor notes generator
